Playable Mei Classroom (w/o Cutscene)
I've implemented a working mirror in the classroom which will be used for the classroom betrayal cutscene by Mei's best friend, Chihaya.

The classroom has also been set up with desks and chairs, alongside Mei's desk with the mourning flower on top of it.

The next thing to do is to create the cutscene and add dialogue interactions to Mei's desk.