Add InteractionManager and LockedInteractable Unique Identifier
A unique identifier is added to the InteractionManager and LockedInteractable scripts. These unique IDs are used to check whether or not certain behaviors should be allowed when loading the same scenes again upon interacting with the object.

This is to ensure player's actions with the objects having these scripts are saved across scenes.

The unique IDs will be saved into a hash set maintained by the Game Manager for efficiency reasons.

The unique IDs are in the format of Scene Name followed by the Game Object name. Since game objects can share the same names across different scenes, the scene name itself is also considered as part of the unique ID.

The unique IDs are defined automatically.