Make Mei Classroom Cutscene (In Progress)
The Mei Classroom cutscene is underway and it consists of the following flow:
- Slowly zoom into the mirror.
- Jump scare by a twisted-looking Mei in the mirror.
- Suddenly zoom out.
- Shows Mei being surrounded by Chihaya and her classmates, clapping and laughing at her.
- Everyone suddenly disappears and the audio stops playing.
- A note remains on the table, with Chihaya and her friends' comments towards Mei on it.

The sudden appearance of Chihaya and the others with the loud SFX is to create a scare effect and build up to the weird things happening in the school.

Let's discuss about this cutscene flow, and how we should proceed to the toilet scene (where dark Mei makes her 2nd appearance after the jump scare.)

We could make the transition into the dark school after the zoom out in the classroom or after leaving the toilet in scene 4.