Add lighting, library and washroom scenes
The renderer in graphics settings is set to Universal Rendering
Pipeline, which allows for lighting effects such as global and
point lightings.

This achieves the desired vibes/ feel for every scene by adjusting
the colors of the lightings during normal and after dark.

I have also rearranged the prefabs into folders so they can be more
easily found. For example, prefabs that are used in every scene can
be found in SceneMustHaves folder so we can just drag all of them
into a new scene.

There is an issue with the mirror effect in the MeiClassroom scene
where the material the mirror sprite has currently does not show the
mirrored texture as intended.

Let's find a solution for this.