Fix animation bug and add new tilemaps and light sprites to scenes
Fixed a bug where the interaction animation is still playing even
when the player is able to move again. This is done by deactivating
the current interaction animation. Then, forcing the animator to go
back to the idle state.

That way the player will always snap back to the idle animation
before being able to move again.

Decorated the home, 1st floor, washroom, and staffroom scenes with
new tilemap sprites and replaced the light sources with new lamp
sprites. Also adjusted how the lighting looks by adjusting the pivot
points of the sprites and the light spread angles.